+++ frontend/src/views/business/reservations/BusinessReservationsPage.vue
@@ -335,46 +335,6 @@ const selectDay = (day) => {
   selectedDate.value = new Date(currentMonth.value.getFullYear(), currentMonth.value.getMonth(), day);
 };
 
-const isWeeklyReportLoading = ref(false);
-
-const downloadWeeklyReport = async () => {
-  const rid = await ensureRestaurantId();
-  if (!rid) return alert("권한이 없습니다.");
-
-  const token = localStorage.getItem('accessToken');
-  if (!token) {
-    window.alert('로그인이 필요합니다.');
-    return;
-  }
-
-  if(userRole.value === 'ROLE_STAFF') return alert("사업자만 요약서 확인이 가능합니다.");
-
-  if (isWeeklyReportLoading.value) return;
-  isWeeklyReportLoading.value = true;
-
-  try {
-    const response = await httpRequest.get(
-        `/api/business/restaurants/${rid}/stats/weekly.pdf`,
-        null,
-        { responseType: 'blob' }
-    );
-
-    const blob = response.data;
-    const url = URL.createObjectURL(blob);
-    const anchor = document.createElement('a');
-    anchor.href = url;
-    anchor.download = `LunchGo-weekly-stats-${rid}.pdf`;
-    document.body.appendChild(anchor);
-    anchor.click();
-    anchor.remove();
-    URL.revokeObjectURL(url);
-  } catch (error) {
-    const status = error?.response?.status ?? 'unknown';
-    window.alert(`문제가 발생했습니다. code: ${status}`);
-  } finally {
-    isWeeklyReportLoading.value = false;
-  }
-};
 
 const settlementChartData = computed(() => {
   if (!settlement.value?.daily) return null;
@@ -502,21 +462,6 @@ const settlementChartOptions = {
         <div class="max-w-7xl mx-auto space-y-8">
           <div class="flex flex-wrap items-center justify-between gap-4">
             <h2 class="text-3xl font-bold text-[#1e3a5f]">전체 예약 관리</h2>
-            <button
-                type="button"
-                @click="downloadWeeklyReport"
-                :disabled="isWeeklyReportLoading"
-                :class="[
-                  'px-4 py-2 rounded-lg text-sm font-semibold text-white bg-gradient-to-r from-[#6366F1] via-[#EC4899] to-[#F97316] transition-opacity',
-                  isWeeklyReportLoading ? 'opacity-70 cursor-not-allowed' : 'hover:opacity-90 cursor-pointer',
-                ]"
-            >
-              <span v-if="isWeeklyReportLoading" class="inline-flex items-center gap-2">
-                <span class="loading-spinner"></span>
-                AI 요약 생성 중...
-              </span>
-              <span v-else>주간 예약 통계 AI 요약 분석서</span>
-            </button>
           </div>
         </div>
 
@@ -818,20 +763,3 @@ const settlementChartOptions = {
     </div>
   </div>
 </template>
-
-<style scoped>
-.loading-spinner {
-  width: 14px;
-  height: 14px;
-  border-radius: 999px;
-  border: 2px solid rgba(255, 255, 255, 0.35);
-  border-top-color: #ffffff;
-  animation: spin 0.8s linear infinite;
-}
-
-@keyframes spin {
-  to {
-    transform: rotate(360deg);
-  }
-}
-</style>
